refactor: agent configuration sheet
