+++ packages/features/calAIPhone/providers/retellAI/types.ts
@@ -113,6 +113,7 @@ export type RetellAgentWithDetails = {
   enabled: boolean;
   userId: number | null;
   teamId: number | null;
+  eventTypeId?: number | null;
   outboundPhoneNumbers: Array<{
     id: number;
     phoneNumber: string;
