+++ packages/features/ee/workflows/components/agent-configuration/utils/constants.ts
@@ -1,0 +1,38 @@
+export const LANGUAGE_OPTIONS = [
+  { value: "en-US", label: "English (US)" },
+  { value: "en-IN", label: "English (India)" },
+  { value: "en-GB", label: "English (UK)" },
+  { value: "en-AU", label: "English (Australia)" },
+  { value: "en-NZ", label: "English (New Zealand)" },
+  { value: "de-DE", label: "German" },
+  { value: "es-ES", label: "Spanish (Spain)" },
+  { value: "es-419", label: "Spanish (Latin America)" },
+  { value: "hi-IN", label: "Hindi" },
+  { value: "fr-FR", label: "French (France)" },
+  { value: "fr-CA", label: "French (Canada)" },
+  { value: "ja-JP", label: "Japanese" },
+  { value: "pt-PT", label: "Portuguese (Portugal)" },
+  { value: "pt-BR", label: "Portuguese (Brazil)" },
+  { value: "zh-CN", label: "Chinese (Simplified)" },
+  { value: "ru-RU", label: "Russian" },
+  { value: "it-IT", label: "Italian" },
+  { value: "ko-KR", label: "Korean" },
+  { value: "nl-NL", label: "Dutch (Netherlands)" },
+  { value: "nl-BE", label: "Dutch (Belgium)" },
+  { value: "pl-PL", label: "Polish" },
+  { value: "tr-TR", label: "Turkish" },
+  { value: "th-TH", label: "Thai" },
+  { value: "vi-VN", label: "Vietnamese" },
+  { value: "ro-RO", label: "Romanian" },
+  { value: "bg-BG", label: "Bulgarian" },
+  { value: "ca-ES", label: "Catalan" },
+  { value: "da-DK", label: "Danish" },
+  { value: "fi-FI", label: "Finnish" },
+  { value: "el-GR", label: "Greek" },
+  { value: "hu-HU", label: "Hungarian" },
+  { value: "id-ID", label: "Indonesian" },
+  { value: "no-NO", label: "Norwegian" },
+  { value: "sk-SK", label: "Slovak" },
+  { value: "sv-SE", label: "Swedish" },
+  { value: "multi", label: "Multilingual" },
+];
