+++ packages/features/ee/workflows/pages/workflow.tsx
@@ -38,7 +38,11 @@ import { getTranslatedText, translateVariablesToEnglish } from "../lib/variableT
 export type FormValues = {
   name: string;
   activeOn: Option[];
-  steps: (WorkflowStep & { senderName: string | null; agentId?: string | null; inboundAgentId?: string | null })[];
+  steps: (WorkflowStep & {
+    senderName: string | null;
+    agentId?: string | null;
+    inboundAgentId?: string | null;
+  })[];
   trigger: WorkflowTriggerEvents;
   time?: number;
   timeUnit?: TimeUnit;
