+++ packages/lib/server/repository/PrismaAgentRepository.ts
@@ -9,6 +9,7 @@ interface _AgentRawResult {
   enabled: boolean;
   userId: number;
   teamId: number | null;
+  eventTypeId?: number | null;
   createdAt: Date;
   updatedAt: Date;
   user_id?: number;
@@ -357,6 +358,7 @@ export class PrismaAgentRepository {
         a.enabled,
         a."userId",
         a."teamId",
+        a."eventTypeId",
         a."createdAt",
         a."updatedAt",
         u.id as user_id,
@@ -398,6 +400,7 @@ export class PrismaAgentRepository {
       enabled: agent.enabled,
       userId: agent.userId,
       teamId: agent.teamId,
+      eventTypeId: agent.eventTypeId,
       createdAt: agent.createdAt,
       updatedAt: agent.updatedAt,
       user: agent.user_id
@@ -565,6 +568,17 @@ export class PrismaAgentRepository {
       },
       data: {
         inboundAgentId: agentId,
+      },
+    });
+  }
+
+  static async updateEventTypeId({ agentId, eventTypeId }: { agentId: string; eventTypeId: number }) {
+    return await prisma.agent.update({
+      where: {
+        id: agentId,
+      },
+      data: {
+        eventTypeId,
       },
     });
   }
